Initial upload from personal computer
This is the 6th iteration of DeuteRater-H from the computer it was developed on.  There are updates to extractor, settings, which statistics to report and so on.  But all major portions, the recursive fit, the error minimization, the calculus are all in places.  As is all the user interaction, settings, and getting the data from the mass spectrometer.